requirements.txt
djangorestframework~=3.12.2
Django~=3.1.6
oauthlib~=3.1.0